backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = ""
authors = [
    { name = "mostafawzy", email = "Mostafawzy33@gmail.com" }
]
readme = "README.md"
requires-python =">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic[dotenv] (>=2.12.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "transformers (>=4.57.0,<5.0.0)",
    "torch (>=2.8.0,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

